javadoc added (#2)
CRUD implementation + FE

+++ src/main/java/org/example/recipesworking/service/implementation/ArticleServiceImp.java
@@ -1,6 +1,8 @@
 package org.example.recipesworking.service.implementation;
 
 import lombok.RequiredArgsConstructor;
+import org.example.recipesworking.exceptions.ArticleGramsOutOfBoundsException;
+import org.example.recipesworking.exceptions.ArticleNotFoundException;
 import org.example.recipesworking.model.Article;
 import org.example.recipesworking.model.dto.ArticleDto;
 import org.example.recipesworking.model.rcord.ArticleRecord;
@@ -36,24 +38,49 @@ public class ArticleServiceImp implements ArticleService {
 
     @Override
     public Article createArticle(ArticleRecord articleRecord) {
-        Article article = new Article(
-                articleRecord.name(),
-                articleRecord.amountInGramPerArticle(),
-                articleRecord.quantity(),
-                articleRecord.caloriesPerGram());
+        boolean articleIsPresent = articleRepository.findByNameAndCaloriesPerGram(articleRecord.name(), articleRecord.caloriesPerGram()).isPresent();
 
-        return articleRepository.save(article);
+        if (articleIsPresent){
+            Article article = articleRepository.findByNameAndCaloriesPerGram(articleRecord.name(), articleRecord.caloriesPerGram()).get();
+            return raiseQuantityOfArticleGrams(article, articleRecord.amountInGramPerArticle() * articleRecord.quantity());
+        }
+            Article article = new Article(
+                    articleRecord.name(),
+                    articleRecord.amountInGramPerArticle(),
+                    articleRecord.quantity(),
+                    articleRecord.caloriesPerGram());
+
+            return articleRepository.save(article);
     }
 
     @Override
     public void deleteArticle(Long articleId) {
-        articleRepository.deleteById(articleId);
+        boolean articleIsPresent = articleRepository.findById(articleId).isPresent();
+        if (articleIsPresent){
+            articleRepository.deleteById(articleId);
+        }
+        throw new ArticleNotFoundException(String.format("Article with id: %s was not found", articleId));
     }
 
     @Override
-    public Article updateArticleGrams(Long articleId, Integer gramsToBeRemoved){
+    public Article removeArticleGrams(Long articleId, Integer gramsToBeRemoved) {
         Article article = articleRepository.getReferenceById(articleId);
-        article.setAmountInGram(article.getAmountInGram() - gramsToBeRemoved);
-        return articleRepository.save(article);
+        int articleGrams = article.getAmountInGram();
+        if (gramsToBeRemoved <= articleGrams) {
+            article.setAmountInGram(article.getAmountInGram() - gramsToBeRemoved);
+            return articleRepository.save(article);
+        }
+        throw new ArticleGramsOutOfBoundsException(String.format("Actual grams %d but to be removed was %d", article.getAmountInGram(), gramsToBeRemoved));
+    }
+
+    @Override
+    public Article raiseQuantityOfArticleGrams(Article article, Integer gramsToBeRaised) {
+            article.setAmountInGram(article.getAmountInGram() + gramsToBeRaised);
+            return articleRepository.save(article);
+    }
+
+    @Override
+    public List<Article> getAllArticles() {
+        return articleRepository.findAll();
     }
 }
